requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
flask-api-key==0.1.7
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
blinker==1.6.2
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
coverage==7.2.3
Deprecated==1.2.13
flake8==6.0.0
Flask==2.3.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Seeder==1.2.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.4.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mccabe==0.7.0
mdurl==0.1.2
migrate==0.3.8
multidict==6.0.4
mypy-extensions==1.0.0
ordered-set==4.1.0
packaging==23.1
pathspec==0.11.1
pip-review==1.3.0
platformdirs==3.4.0
psycopg2-binary==2.9.10
psycopg2==2.9.10
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
regex==2023.3.23
rich==13.3.4
rq==1.13.0
six==1.16.0
SQLAlchemy==2.0.10
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.0
Werkzeug==2.3.0
wrapt==1.15.0
yarl==1.9.2
